probabilistic-predictions/requirements.txt
pandas~=2.2.3
numpy==1.26.4
neo4j~=5.26.0
pykeen
python-dotenv~=1.0.1
torch~=2.5.1
pyro-ppl
matplotlib~=3.9.2